requirements.txt
absl-py
accelerate
aiobotocore
aiohappyeyeballs
aiohttp
aioitertools
aiosignal
alembic
annotated-types
anyio
argon2-cffi
argon2-cffi-bindings
arrow
astropy
astropy-iers-data
asttokens
astunparse
async-lru
asyncio
attrs
babel
beautifulsoup4
bleach
botocore
cachetools
certifi
cffi
charset-normalizer
click
cloudpickle
colorama
colorlog
comm
contourpy
cramjam
crc32c
cycler
datasets
debugpy
decorator
defusedxml
dill
donfig
einops
evaluate
executing
Farama-Notifications
fastapi
fastjsonschema
fastparquet
filelock
flatbuffers
fonttools
fqdn
frozenlist
fsspec
gast
gitdb
GitPython
google-pasta
greenlet
grpcio
gymnasium
h11
h5py
hnswlib
httpcore
httpx
huggingface-hub
idna
isoduration
jedi
Jinja2
jmespath
joblib
json5
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
kaggle
keras
kiwisolver
langchain-core
langsmith
lark
libclang
Mako
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
mistune
ml_dtypes
mpmath
multidict
multiprocess
namex
nest-asyncio
networkx
numcodecs
numpy
onnx
opt_einsum
optimum
optree
optuna
orjson
overrides
packaging
pandas
pandocfilters
parso
peft
pillow
platformdirs
prometheus_client
prompt_toolkit
propcache
protobuf
psutil
pure_eval
pyarrow
pycparser
pydantic
pydantic_core
pyerfa
pyflakes
Pygments
pyparsing
python-dateutil
python-dotenv
python-json-logger
python-slugify
pytz
pywin32-ctypes
PyYAML
pyzmq
referencing
regex
requests
requests-toolbelt
rfc3339-validator
rfc3986-validator
rich
rpds-py
s3transfer
scikit-learn
scipy
Send2Trash
six
smmap
sniffio
soupsieve
SQLAlchemy
stack_data
starlette
sympy
tblib
tenacity
tensorboard
tensorboard-data-server
tensorflow
tensorflow-io-gcs-filesystem
termcolor
terminado
text-unidecode
threadpoolctl
tinycss2
tokenizers
tiktoken
toml
tomlkit
toolz
torch
torchaudio
torchvision
tornado
tqdm
traitlets
transformers
types-python-dateutil
typing_extensions
tzdata
uri-template
urllib3
uvicorn
wandb
wcwidth
webcolors
webencodings
websocket-client
Werkzeug
wrapt
xxhash
yarl
zarr
zipp
kafka-python
langchain
langgraph
redis
aiokafka
watchfiles
websockets
zipp
nvidia-physicsnemo
pandas
openai
anthropic
google-generativeai
google-genai
google-cloud-aiplatform
#modulus-sym
# Temporary comment to invalidate cache
